homework_06/hasker/requirements.txt
pylint==2.17.4
pylint-runner==0.6.0
pytest==7.4.2
python-dateutil==2.8.2
redis==5.0.1
pytest-xdist==3.3.1
Django==4.2.6
Pillow==10.1.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0